refactor: 내부에 혼재되어 있던 타입 정의들을 별도 파일로 분리
